Испытываю программу на предмет багов.

+++ tg_master.py
@@ -13,7 +13,7 @@ def test(message_):
 
 
 def message(name, id_, new_price, search_price, name_in_search):
-    new_price, search_price = float(new_price), float(search_price)
+    new_price, search_price = int(new_price), int(search_price)
     bot.send_message(674796107, f'Товар: {name}, \n id: {id_} стоимостью - {new_price}  упал в цене. \n'
                              f'https://www.wildberries.ru/catalog/{id_}/detail.aspx'
                              f'\n В Яндекс найден похожий товар: \n'
